+++ pages/events/[filename].tsx
@@ -14,6 +14,7 @@ import { Container } from "../../components/util/container";
 import { Section } from "../../components/util/section";
 import { SEO } from "../../components/util/seo";
 import VideoCards, { VideoCardProps } from "../../components/util/videoCards";
+import { GetTestimonialsByCategories } from "../../helpers/getTestimonials";
 import { removeExtension } from "../../services/client/utils.service";
 
 export default function EventsPage(
@@ -121,12 +122,14 @@ export const getStaticProps = async ({ params }) => {
     relativePath: `${params.filename}.mdx`,
   });
 
-  const testimonials = await client.queries.testimonalsQuery({
-    categories: "Internship",
-  });
+  const categories =
+    tinaProps.data.events?.testimonialCategories?.map(
+      (category) => category.testimonialCategory.name
+    ) || [];
 
-  const testimonialsResult = testimonials.data.testimonialsConnection.edges.map(
-    (t) => t.node
+  const testimonialsResult = await GetTestimonialsByCategories(
+    categories,
+    "Internship"
   );
 
   if (tinaProps.data.events.seo && !tinaProps.data.events.seo.canonical) {
